+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.1] - 2024-08-09
+
 ### Added
 
 #### Position argument parsing
@@ -29,5 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Build system.
 
-[Unreleased]: https://github.com/sonro/spat/compare/v0.0.0...HEAD
-[0.0.0]: https://github.com/sonro/zaplum/releases/tag/v0.0.0
+[Unreleased]: https://github.com/sonro/spat/compare/v0.0.1...HEAD
+[0.0.1]: https://github.com/sonro/spat/releases/tag/v0.0.1
+[0.0.0]: https://github.com/sonro/spat/releases/tag/v0.0.0